# docs/fern/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SeatSure API",
    "version": "0.1.0",
    "description": "HTTP API of SeatSure. Every success response is `{ \"success\": true, \"data\": ... }`; every failure is `{ \"success\": false, \"message\": \"<code>\", \"errors\"?: [...] }`. Generated from the endpoint registry; do not edit by hand."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development (`task up`)"
    }
  ],
...
  "paths": {
...
    "/bookings": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "operationId": "bookSeat",
        "summary": "Book a seat in a course",
        "description": "Holds one seat for the named student, as the signed-in user, and answers `201` with the new booking (status `held`). The seat count is checked under a row lock, so concurrent requests for the last seat give exactly one booking and `course_full` to the rest. An account can have one held or paid booking per course: a second request (for example a double-click) answers `already_booked`. Holds that ran out are released first. A transfer booking stays held for 7 days until payment is confirmed. The student name is trimmed; an empty or blank `studentName` answers `student_name_required` (the server checks it before the schema's non-empty rule, so it is never a `Validation failed`).\n\nRequires a bearer token (any role).\n\nErrors (the `message` of the failure envelope):\n\n- `400` `student_name_required`: The student name is empty or blank.\n- `400` `Validation failed`: `courseId` is missing or not a uuid; see `errors`.\n- `400` `course_full`: Every seat is paid or held.\n- `400` `registration_closed`: The course does not accept bookings now.\n- `400` `already_booked`: This account already holds or paid a seat in the course.\n- `400` `course_not_approved`: The course is pending, rejected or cancelled.\n- `401` `not_authenticated`: Missing, invalid or expired bearer token.\n- `404` `course_not_found`",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookSeatBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/Booking"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "`student_name_required`, `Validation failed`, `course_full`, `registration_closed`, `already_booked`, `course_not_approved`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            }
          },
          "401": {
            "description": "`not_authenticated`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            }
          },
          "404": {
            "description": "`course_not_found`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/bookings/mine": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "operationId": "listMyBookings",
        "summary": "List my bookings",
        "description": "The signed-in user's own bookings, newest first, each with the course title and price, its payments (receipts and `refund_due` entries) and its transfer proof. Every role sees only its own bookings here.\n\nRequires a bearer token (any role).\n\nErrors (the `message` of the failure envelope):\n\n- `401` `not_authenticated`: Missing, invalid or expired bearer token.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Booking"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "`not_authenticated`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/bookings/active": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "operationId": "listActiveBookings",
        "summary": "List held and paid bookings I can see",
        "description": "Bookings with status `held` or `paid`, oldest first, without joins. A parent gets their own; a teacher gets their own plus the bookings of the courses they teach; an admin gets all. A `held` row whose `hold_expires_at` has passed no longer holds a seat: the client compares it with the current time (paid, or held and not yet expired, holds a seat). Used by the course cards (\"already booked\") and the teacher's student lists.\n\nRequires a bearer token (any role).\n\nErrors (the `message` of the failure envelope):\n\n- `401` `not_authenticated`: Missing, invalid or expired bearer token.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ActiveBooking"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "`not_authenticated`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/bookings/{id}": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "operationId": "getBooking",
        "summary": "Get one of my bookings",
        "description": "One booking of the signed-in user with course, payments and transfer proof (the receipt page). Bookings of other accounts, unknown ids and malformed ids all answer `booking_not_found`, so the API does not reveal whether a booking exists.\n\nRequires a bearer token (any role).\n\nErrors (the `message` of the failure envelope):\n\n- `401` `not_authenticated`: Missing, invalid or expired bearer token.\n- `404` `booking_not_found`: No booking with this id belongs to you.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/Booking"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "`not_authenticated`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            }
          },
          "404": {
            "description": "`booking_not_found`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/courses/{id}/roster": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "operationId": "getCourseRoster",
        "summary": "List the students of a course",
        "description": "Every booking of the course (any status), oldest first. Allowed for the course's teacher and for admins; another teacher gets `403 forbidden` and no data. Each viewer sees what the database policies allow them: the teacher gets the booking rows and student names only (`profiles.full_name` is `\"\"`, `payments` and `payment_proofs` are empty); an admin also gets the parent's name and the payments; only the school admin (`admin01@seatsure.test`) also gets the transfer proofs, each with a `signed_url` to the private image valid for 600 seconds. For any other admin `payment_proofs` is empty.\n\nRequires a bearer token with role: teacher, admin.\n\nErrors (the `message` of the failure envelope):\n\n- `401` `not_authenticated`: Missing, invalid or expired bearer token.\n- `403` `forbidden`: Not an admin, and not the teacher of this course.\n- `404` `course_not_found`",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/RosterRow"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "`not_authenticated`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            }
          },
          "404": {
            "description": "`course_not_found`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "The `accessToken` from `POST /auth/login` or `POST /auth/refresh`."
      }
    },
    "schemas": {
      "ActiveBooking": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "course_id": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "hold_expires_at": {
            "type": "string"
          },
          "paid_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "held",
              "paid",
              "cancelled",
              "expired"
            ]
          },
          "student_name": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "course_id",
          "created_at",
          "hold_expires_at",
          "paid_at",
          "status",
          "student_name",
          "user_id"
        ],
        "additionalProperties": false
      },
...
      "Booking": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "course_id": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "hold_expires_at": {
            "type": "string"
          },
          "paid_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "held",
              "paid",
              "cancelled",
              "expired"
            ]
          },
          "student_name": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "courses": {
            "type": "object",
            "properties": {
              "title": {
                "type": "string"
              },
              "price": {
                "type": "number"
              }
            },
            "required": [
              "title",
              "price"
            ],
            "additionalProperties": false
          },
          "payments": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "booking_id": {
                  "type": "string"
                },
                "amount": {
                  "type": "number"
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "succeeded",
                    "refund_due"
                  ]
                },
                "receipt_no": {
                  "type": "string"
                },
                "created_at": {
                  "type": "string"
                },
                "idempotency_key": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "booking_id",
                "amount",
                "status",
                "receipt_no",
                "created_at",
                "idempotency_key"
              ],
              "additionalProperties": false
            }
          },
          "payment_proofs": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "booking_id": {
                  "type": "string"
                },
                "proof_path": {
                  "type": "string"
                },
                "submitted_at": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "booking_id",
                "proof_path",
                "submitted_at"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "id",
          "course_id",
          "created_at",
          "hold_expires_at",
          "paid_at",
          "status",
          "student_name",
          "user_id",
          "courses",
          "payments",
          "payment_proofs"
        ],
        "additionalProperties": false
      },
      "BookSeatBody": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "studentName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        },
        "required": [
          "courseId",
          "studentName"
        ]
      },
...
      "Failure": {
        "type": "object",
        "description": "Every error leaves the API in this envelope. `message` is the error code (or `Validation failed`).",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
      "FieldError": {
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "RosterRow": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "student_name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "held",
              "paid",
              "cancelled",
              "expired"
            ]
          },
          "hold_expires_at": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "profiles": {
            "type": "object",
            "properties": {
              "full_name": {
                "type": "string"
              }
            },
            "required": [
              "full_name"
            ],
            "additionalProperties": false
          },
          "payments": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "amount": {
                  "type": "number"
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "succeeded",
                    "refund_due"
                  ]
                },
                "receipt_no": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "amount",
                "status",
                "receipt_no"
              ],
              "additionalProperties": false
            }
          },
          "payment_proofs": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "booking_id": {
                  "type": "string"
                },
                "proof_path": {
                  "type": "string"
                },
                "submitted_at": {
                  "type": "string"
                },
                "signed_url": {
                  "type": [
                    "string",
                    "null"
                  ]
                }
              },
              "required": [
                "id",
                "booking_id",
                "proof_path",
                "submitted_at",
                "signed_url"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "id",
          "student_name",
          "status",
          "hold_expires_at",
          "created_at",
          "profiles",
          "payments",
          "payment_proofs"
        ],
        "additionalProperties": false
      },
...
    }
  }
}